add aria label to item wrapper
